+++ src/engine/script/reflection.h
@@ -192,8 +192,8 @@ public:
                     bool flipRooms,
                     bool weaponSwap,
                     float cameraRot,
-                    int cameraPosX,
-                    int cameraPosZ)
+                    float cameraPosX,
+                    float cameraPosZ)
       : m_name{std::move(name)}
       , m_track{track}
       , m_flipRooms{flipRooms}
